+++ tellez.aron/I4/matrixWaveValue.h
@@ -1,5 +1,5 @@
 #ifndef SPBSPU_LABS_2022_AAP_904_A_MATRIXWAVEVALUE_HPP
 #define SPBSPU_LABS_2022_AAP_904_A_MATRIXWAVEVALUE_HPP
 #include <cstddef>
-int* assignMatrixWaveValue(int *matrix, size_t cols, size_t rows, size_t index_row, size_t index_col);
+size_t* assignMatrixWaveValue(size_t *matrix, size_t cols, size_t rows, size_t index_row, size_t index_col);
 #endif
